# numguess-service/src/main/java/org/openapitools/api/GameApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Error;
import org.openapitools.model.GameState;
import org.openapitools.model.GuessResult;
import org.springframework.lang.Nullable;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-10T20:40:10.675029-06:00[America/Chicago]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "game", description = "Operations for individual game instances")
public interface GameApi {

    default GameApiDelegate getDelegate() {
        return new GameApiDelegate() {};
    }

    String PATH_GAMES_UUID_DELETE = "/games/{uuid}";
    /**
     * DELETE /games/{uuid} : Delete a game instance
     * Deletes the specified game instance identified by its UUID. This removes the game from the collection of active games.  Following REST principles, this operation is idempotent - deleting a non-existent game should return 404, but deleting the same game multiple times has the same effect as deleting it once. 
     *
     * @param uuid The unique identifier (UUID) of the game to delete (required)
     * @return Game deleted successfully (No Content) (status code 204)
     *         or Game not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "gamesUuidDelete",
        summary = "Delete a game instance",
        description = "Deletes the specified game instance identified by its UUID. This removes the game from the collection of active games.  Following REST principles, this operation is idempotent - deleting a non-existent game should return 404, but deleting the same game multiple times has the same effect as deleting it once. ",
        tags = { "game" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Game deleted successfully (No Content)"),
            @ApiResponse(responseCode = "404", description = "Game not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = GameApi.PATH_GAMES_UUID_DELETE,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> gamesUuidDelete(
        @NotNull @Parameter(name = "uuid", description = "The unique identifier (UUID) of the game to delete", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid
    ) {
        return getDelegate().gamesUuidDelete(uuid);
    }


    String PATH_GAMES_UUID_GET = "/games/{uuid}";
    /**
     * GET /games/{uuid} : Get game representation
     * Returns a form (HTML) or game state representation with hypermedia controls (JSON). The game allows guessing numbers between 1 and 100.  For JSON clients, provides full hypermedia controls including: - Link to submit a guess (if game is still active) - Link to delete the game - Navigation links to related resources - Current game state information  This enables client discoverability of available actions (Level 3 HATEOAS). 
     *
     * @param uuid The unique identifier (UUID) of the game (required)
     * @param accept Content type preference (optional, default to application/xhtml+xml)
     * @return Game representation with hypermedia controls (status code 200)
     *         or Game not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "gamesUuidGet",
        summary = "Get game representation",
        description = "Returns a form (HTML) or game state representation with hypermedia controls (JSON). The game allows guessing numbers between 1 and 100.  For JSON clients, provides full hypermedia controls including: - Link to submit a guess (if game is still active) - Link to delete the game - Navigation links to related resources - Current game state information  This enables client discoverability of available actions (Level 3 HATEOAS). ",
        tags = { "game" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Game representation with hypermedia controls", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "404", description = "Game not found", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = GameApi.PATH_GAMES_UUID_GET,
        produces = { "application/xhtml+xml", "application/json", "application/hal+json" }
    )
    default ResponseEntity<String> gamesUuidGet(
        @NotNull @Parameter(name = "uuid", description = "The unique identifier (UUID) of the game", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid,
        @Parameter(name = "Accept", description = "Content type preference", in = ParameterIn.HEADER) @RequestHeader(value = "Accept", required = false, defaultValue = "application/xhtml+xml") String accept
    ) {
        return getDelegate().gamesUuidGet(uuid, accept);
    }


    String PATH_GAMES_UUID_POST = "/games/{uuid}";
    /**
     * POST /games/{uuid} : Submit a guess
     * Submits a guess for the specified game. The guess is compared against the secret number, and a result is returned indicating whether the guess was correct, too high, or too low.  The response includes comprehensive hypermedia controls (Level 3 HATEOAS): - Current game state and guess result - Contextual action links based on game state:   * If game is still active: link to submit another guess   * If game is won: no submit-guess link (demonstrates state-driven affordances) - Navigation links to create new games or return to collection - Number of guesses made so far - Best score tracking across all games  The hypermedia controls enable clients to navigate the application state without hardcoding URLs or business logic about when actions are available. 
     *
     * @param uuid The unique identifier (UUID) of the game (required)
     * @param guess The guessed number (must be between 1 and 100) (required)
     * @param accept Content type preference (optional, default to application/xhtml+xml)
     * @return Guess processed successfully with hypermedia controls (status code 201)
     *         or Game not found (status code 404)
     *         or Bad request (invalid guess format or value) (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "gamesUuidPost",
        summary = "Submit a guess",
        description = "Submits a guess for the specified game. The guess is compared against the secret number, and a result is returned indicating whether the guess was correct, too high, or too low.  The response includes comprehensive hypermedia controls (Level 3 HATEOAS): - Current game state and guess result - Contextual action links based on game state:   * If game is still active: link to submit another guess   * If game is won: no submit-guess link (demonstrates state-driven affordances) - Navigation links to create new games or return to collection - Number of guesses made so far - Best score tracking across all games  The hypermedia controls enable clients to navigate the application state without hardcoding URLs or business logic about when actions are available. ",
        tags = { "game" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Guess processed successfully with hypermedia controls", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = String.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"gameId\":\"550e8400-e29b-41d4-a716-446655440000\",\"guess\":42,\"result\":\"correct\",\"message\":\"Congratulations! You guessed the correct number in 7 tries!\",\"numGuesses\":7,\"bestScore\":7,\"newBestScore\":true,\"_links\":{\"self\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"type\":\"application/json\"},\"delete\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"method\":\"DELETE\",\"title\":\"Delete this game\"},\"new-game\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"method\":\"POST\",\"type\":\"application/json\",\"title\":\"Create a new game\"},\"games\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"type\":\"application/json\",\"title\":\"Games collection\"}}}"
                    ),
                    @ExampleObject(
                        name = "",
                        value = "{\"gameId\":\"550e8400-e29b-41d4-a716-446655440000\",\"guess\":25,\"result\":\"too_low\",\"message\":\"Your guess is too low. Try a higher number.\",\"numGuesses\":3,\"bestScore\":5,\"newBestScore\":false,\"_links\":{\"self\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"type\":\"application/json\"},\"submit-guess\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"method\":\"POST\",\"type\":\"application/x-www-form-urlencoded\",\"title\":\"Submit another guess\"},\"delete\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"method\":\"DELETE\",\"title\":\"Delete this game\"},\"new-game\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"method\":\"POST\",\"type\":\"application/json\",\"title\":\"Create a new game\"},\"games\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"type\":\"application/json\",\"title\":\"Games collection\"}}}"
                    ),
                    @ExampleObject(
                        name = "",
                        value = "{\"gameId\":\"550e8400-e29b-41d4-a716-446655440000\",\"guess\":75,\"result\":\"too_high\",\"message\":\"Your guess is too high. Try a lower number.\",\"numGuesses\":2,\"bestScore\":5,\"newBestScore\":false,\"_links\":{\"self\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"type\":\"application/json\"},\"submit-guess\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"method\":\"POST\",\"type\":\"application/x-www-form-urlencoded\",\"title\":\"Submit another guess\"},\"delete\":{\"href\":\"http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000\",\"method\":\"DELETE\",\"title\":\"Delete this game\"},\"new-game\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"method\":\"POST\",\"type\":\"application/json\",\"title\":\"Create a new game\"},\"games\":{\"href\":\"http://localhost:3000/numguess-restlet/games\",\"type\":\"application/json\",\"title\":\"Games collection\"}}}"
                    )
                }),
                
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "404", description = "Game not found", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request (invalid guess format or value)", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/xhtml+xml", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/hal+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = GameApi.PATH_GAMES_UUID_POST,
        produces = { "application/xhtml+xml", "application/json", "application/hal+json" },
        consumes = { "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<String> gamesUuidPost(
        @NotNull @Parameter(name = "uuid", description = "The unique identifier (UUID) of the game", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid,
        @Parameter(name = "guess", description = "The guessed number (must be between 1 and 100)", required = true) @Min(value = 1) @Max(value = 100) @Valid @RequestParam(value = "guess", required = true) Integer guess,
        @Parameter(name = "Accept", description = "Content type preference", in = ParameterIn.HEADER) @RequestHeader(value = "Accept", required = false, defaultValue = "application/xhtml+xml") String accept
    ) {
        return getDelegate().gamesUuidPost(uuid, guess, accept);
    }

}